Add --rust-target to replace --unstable-rust
Instead of specifying whether or not to use stable, specify the Rust
release to support (one of several stable/beta releases or nightly).
The --unstable-rust option is still accepted and implies nightly.

The definitions of `RustTarget` and `RustFeatures` are created with
macros.

In order to keep the test outputs the same,
`bindgen-flags: --rust-target 1.0` was added to test headers.